% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.VEwaningVariant.R
\name{print}
\alias{print}
\alias{print.VEwaningVariant}
\title{Print Analysis Results}
\usage{
\method{print}{VEwaningVariant}(x, ...)
}
\arguments{
\item{x}{An object of class VEwaningVariant. The object returned by a call to
veWaningVariant()}

\item{...}{Ignored}
}
\value{
No return value, called to display key results.
}
\description{
Print the primary results of the analysis
}
\examples{
data(variantData)

set.seed(1234)

ind <- sample(1:nrow(variantData), 2000)
# NOTE: This sample size is chosen for example only -- larger data sets
# should be used.

res <- veWaningVariant(data = variantData[ind,], 
                       L = 52,  
                       gFunc = 'piece', 
                       v = c(15,30))

print(x = res)
}
